another transforms approach

+++ torchgeo/datasets/deepglobelandcover.py
@@ -173,7 +173,7 @@ class DeepGlobeLandCover(NonGeoDataset):
             array: "np.typing.NDArray[np.int_]" = np.array(img)
             tensor = torch.from_numpy(array)
             # Convert from HxWxC to CxHxW
-            tensor = tensor.permute((2, 0, 1))
+            tensor = tensor.permute((2, 0, 1)).to(torch.float32)
             return tensor
 
     def _load_target(self, index: int) -> Tensor:
